Port to SDL3 + CMake for Apple Silicon (arm64)

+++ Engine/src/draw.h
@@ -34,10 +34,10 @@ void sethlinesizes(int32_t,int32_t,uint8_t *);
 
 
 void hlineasm4(int32_t,int32_t,uint32_t,uint32_t,uint8_t*);
-void setuprhlineasm4(int32_t,int32_t,int32_t,int32_t,int32_t,int32_t);
-void rhlineasm4(int32_t,uint8_t*,int32_t,uint32_t,uint32_t,int32_t);
-void setuprmhlineasm4(int32_t,int32_t,int32_t,int32_t,int32_t,int32_t);
-void rmhlineasm4(int32_t,intptr_t,int32_t,int32_t,int32_t,int32_t);
+void setuprhlineasm4(int32_t,int32_t,int32_t,intptr_t,int32_t,int32_t);
+void rhlineasm4(int32_t,uint8_t*,int32_t,uint32_t,uint32_t,intptr_t);
+void setuprmhlineasm4(int32_t,int32_t,int32_t,intptr_t,int32_t,int32_t);
+void rmhlineasm4(int32_t,intptr_t,int32_t,int32_t,int32_t,intptr_t);
 
 
 void setBytesPerLine(int32_t);
@@ -46,16 +46,16 @@ int32_t prevlineasm1(int32_t,uint8_t* palette ,int32_t,int32_t,uint8_t  *,uint8_
 int32_t vlineasm1(int32_t,uint8_t*,int32_t,int32_t,uint8_t  *,uint8_t*);
 
 int32_t tvlineasm1(int32_t,uint8_t  *,int32_t,int32_t,uint8_t  *,uint8_t  * dest);
-void setuptvlineasm2(int32_t,int32_t,int32_t);
+void setuptvlineasm2(int32_t,uintptr_t,uintptr_t);
 void tvlineasm2(uint32_t,uint32_t,uintptr_t,uintptr_t,uint32_t,uintptr_t);
 int32_t mvlineasm1(int32_t,uint8_t*,int32_t,int32_t,uint8_t* texture,uint8_t* dest);
 void setupvlineasm(int32_t);
 void vlineasm4(int32_t,intptr_t);
 void setupmvlineasm(int32_t);
 void mvlineasm4(int32_t,intptr_t);
-void setupspritevline(int32_t,int32_t,int32_t,int32_t,int32_t,int32_t);
+void setupspritevline(intptr_t,int32_t,int32_t,int32_t,int32_t,int32_t);
 void spritevline(int32_t,uint32_t,int32_t,uint32_t,uint8_t*,uint8_t*);
-void msetupspritevline(int32_t,int32_t,int32_t,int32_t,int32_t,int32_t);
+void msetupspritevline(intptr_t,int32_t,int32_t,int32_t,int32_t,int32_t);
 void mspritevline(int32_t,int32_t,int32_t,int32_t,uint8_t  *,uint8_t  *);
 void tsetupspritevline(uint8_t *,int32_t,int32_t,int32_t,int32_t);
 void DrawSpriteVerticalLine(int32_t,int32_t,uint32_t,uint8_t* ,uint8_t*);
@@ -66,7 +66,7 @@ void thline(uint8_t*,int32_t,int32_t,int32_t,int32_t,uint8_t *);
 void thlineskipmodify(int32_t,uint32_t,uint32_t,int32_t,int32_t,uint8_t *);
 void tsethlineshift(int32_t,int32_t);
 void setupslopevlin(int32_t,intptr_t,int32_t);
-void slopevlin(intptr_t,uint32_t,int32_t,int32_t,int32_t,int32_t);
+void slopevlin(intptr_t,uint32_t,intptr_t,int32_t,int32_t,int32_t);
     
     
 #define TRANS_NORMAL  0
